code salary monthly

+++ src/main/java/com/csproject/hrm/dto/response/PointResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class PointResponse {
-  private int actual_working_day;
-  private int unpaid_leave_day;
-  private int paid_leave_day;
+  private Double actual_working_day;
+  private Double unpaid_leave_day;
+  private Double paid_leave_day;
 }
